requirements.txt
Flask>=3.0.0,<4.0.0
gunicorn>=21.2.0,<22.0.0
google-generativeai>=0.7.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
